backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-cors-headers
# Manu
channels
channels-redis
daphne
websockets